+++ docs/release-notes.md
@@ -22,7 +22,8 @@
   * Added prerelease versions to the mod update-check API response where available (GitHub only).
   * Added support for beta releases on the home page.
   * Split mod DB out of `StardewModdingAPI.config.json`, so we can load config earlier and reduce unnecessary memory usage later.
-  * Overhauled world/player state tracking:
+  * Rewrote input suppression using new SDV 1.3 APIs.
+  * Rewrote world/player state tracking:
     * much more efficient than previous method;
     * uses net field events where available;
     * lays groundwork for tracking events for multiple players.